use-cases/05-ai-data-pipeline/requirements.txt
fastapi>=0.135.0
uvicorn>=0.41.0
pydantic>=2.12.0
python-dotenv>=1.2.0
numpy>=1.24.0
httpx>=0.28.0
